Move coach routes to own module
